Fix: 간헐적 테스트 실패 문제 해결 (#137)
* [#134] Fix: BigDecimal 소수점 문제 해결

* [#134] Fix: 도메인 별 fixture에 제약조건 추가

* [#134] Fix: 도메인 별 fixture에 최소 길이의 고정 제약 조건을 추가

+++ src/test/java/soon/fridgely/domain/food/controller/FoodControllerTest.java
@@ -183,7 +183,7 @@ class FoodControllerTest extends ControllerTestSupport {
             .andExpect(jsonPath("$.data.id").value(response.id()))
             .andExpect(jsonPath("$.data.name").value(response.name()))
             .andExpect(jsonPath("$.data.categoryName").value(response.categoryName()))
-            .andExpect(jsonPath("$.data.quantity.amount").value(response.quantity().amount()))
+            .andExpect(jsonPath("$.data.quantity.amount").value(response.quantity().amount().doubleValue()))
             .andExpect(jsonPath("$.data.quantity.unit").value(response.quantity().unit().name()))
             .andExpect(jsonPath("$.data.condition.storageType").value(response.condition().storageType().name()))
             .andExpect(jsonPath("$.data.condition.foodStatus").value(response.condition().foodStatus().name()));
